+++ library/PluginManager.cpp
@@ -24,6 +24,7 @@ distribution.
 
 #include "modules/EventManager.h"
 #include "modules/Filesystem.h"
+#include "modules/Screen.h"
 #include "Internal.h"
 #include "Core.h"
 #include "MemAccess.h"
@@ -272,23 +273,13 @@ bool Plugin::load(color_ostream &con)
             plugin_abort_load; \
             return false; \
         }
-    #define plugin_check_symbols(sym1,sym2) \
-        if (!LookupPlugin(plug, sym1) && !LookupPlugin(plug, sym2)) \
-        { \
-            con.printerr("Plugin %s: missing symbols: %s & %s\n", name.c_str(), sym1, sym2); \
-            plugin_abort_load; \
-            return false; \
-        }
 
-    plugin_check_symbols("plugin_name", "name")                 // allow r3 plugins
-    plugin_check_symbols("plugin_version", "version")           // allow r3 plugins
+    plugin_check_symbol("plugin_name")
+    plugin_check_symbol("plugin_version")
     plugin_check_symbol("plugin_self")
     plugin_check_symbol("plugin_init")
     plugin_check_symbol("plugin_globals")
     const char ** plug_name =(const char ** ) LookupPlugin(plug, "plugin_name");
-    if (!plug_name)                                            // allow r3 plugin naming
-        plug_name = (const char ** )LookupPlugin(plug, "name");
-
     if (name != *plug_name)
     {
         con.printerr("Plugin %s: name mismatch, claims to be %s\n", name.c_str(), *plug_name);
@@ -296,9 +287,6 @@ bool Plugin::load(color_ostream &con)
         return false;
     }
     const char ** plug_version =(const char ** ) LookupPlugin(plug, "plugin_version");
-    if (!plug_version)                                         // allow r3 plugin version
-        plug_version =(const char ** ) LookupPlugin(plug, "version");
-
     const char ** plug_git_desc_ptr = (const char**) LookupPlugin(plug, "plugin_git_description");
     Plugin **plug_self = (Plugin**)LookupPlugin(plug, "plugin_self");
     const char *dfhack_version = Version::dfhack_version();
@@ -385,6 +373,12 @@ bool Plugin::unload(color_ostream &con)
     // if we are actually loaded
     if(state == PS_LOADED)
     {
+        if (Screen::hasActiveScreens(this))
+        {
+            con.printerr("Cannot unload plugin %s: has active viewscreens\n", name.c_str());
+            access->unlock();
+            return false;
+        }
         EventManager::unregisterAll(this);
         // notify the plugin about an attempt to shutdown
         if (plugin_onstatechange &&