+++ library/include/modules/Screen.h
@@ -54,6 +54,7 @@ namespace df
 namespace DFHack
 {
     class Core;
+    class Plugin;
 
     typedef std::set<df::interface_key> interface_key_set;
 
@@ -208,9 +209,12 @@ namespace DFHack
         DFHACK_EXPORT bool findGraphicsTile(const std::string &page, int x, int y, int *ptile, int *pgs = NULL);
 
         // Push and remove viewscreens
-        DFHACK_EXPORT bool show(df::viewscreen *screen, df::viewscreen *before = NULL);
+        DFHACK_EXPORT bool show(df::viewscreen *screen, df::viewscreen *before = NULL, Plugin *p = NULL);
+        inline bool show(df::viewscreen *screen, Plugin *p)
+            { return show(screen, NULL, p); }
         DFHACK_EXPORT void dismiss(df::viewscreen *screen, bool to_first = false);
         DFHACK_EXPORT bool isDismissed(df::viewscreen *screen);
+        DFHACK_EXPORT bool hasActiveScreens(Plugin *p);
 
         /// Retrieve the string representation of the bound key.
         DFHACK_EXPORT std::string getKeyDisplay(df::interface_key key);